requirements.txt
discord.py==1.7.3
Google_Images_Search==1.4.1
Pillow==9.0.1
python-dotenv==0.19.2
